backend/project/asset_management/pyproject.toml
[tool.poetry]
name = "asset-management"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "n"
readme = "README.md"
packages = [{include = "asset_management"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.4"
psycopg2 = "^2.9.6"
django-environ = "^0.10.0"
graphene-django = "^3.1.5"
django-graphql-jwt = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
